refactor files of create user functionality
