+++ src/main/kotlin/logic/usecases/auth/CreateUserUseCase.kt
@@ -1,20 +1,31 @@
 package logic.usecases.auth
 
 import logic.entities.User
+import logic.exception.PlanMateException
 import logic.repository.AuthRepository
-import ui.utils.ResultStatus
+import logic.usecases.validation.ValidateInputUseCase
 import java.util.UUID
 
-class CreateUserUseCase(private val repository: AuthRepository) {
+class CreateUserUseCase(private val repository: AuthRepository, private val validateInputUseCase: ValidateInputUseCase) {
 
-    fun execute(username: String, password: String, isAdmin: Boolean): ResultStatus<User> {
-        val dummyUser = User(
+    fun execute(username: String, password: String): Boolean {
+        if (! validateInputUseCase.isValidName(username)) throw PlanMateException.ValidationException.InvalidUsernameException
+        validatePassword(password)
+        val user = User(
             id = UUID.randomUUID(),
-            username = "",
-            password = "",
+            username = username,
+            password = md5Hash(password),
             isAdmin = false
         )
+        return repository.createUser(user)
 
-        return ResultStatus.Success(dummyUser)
     }
+
+
+    private fun validatePassword(password:String){
+        if (password.isBlank() || password.isEmpty() || password.length < 8){
+            throw PlanMateException.ValidationException.InvalidPasswordException
+        }
+    }
+
 }
